setup scroll to work on any element and add x scrolling
- split up actions test
